pausado API Almacenamiento web
